added naive center and scale implementation and tested. will need to optimize: this is not the best way to solve it, extremely low occupancy with longer col as sh mem fills up: crashes at around 750 rows.

added python code to generate sample input matrices specifically for testing my PCA code. columns of different means and variances, and certain columns correlate with one another (with noise)

also added some helpful strucutre (header w func to test for cuda errors, etc)

See conversation for further todos

+++ centerAndScale.h
@@ -1,6 +1,6 @@
 #ifndef CENTERANDSCALE_H
 #define CENTERANDSCALE_H
 
-int* centerAndScaleWrapper(int* inputMatrixHost, int rowCount, int colCount, bool debugMode);
+float* centerAndScaleWrapper(float* inputMatrixHost, int rowCount, int colCount, bool debugMode);
 
 #endif
